+++ modules/m_dline.c
@@ -168,7 +168,7 @@ dline_handle(struct Client *source, const struct aline_ctx *aline)
 }
 
 static void
-mo_dline(struct Client *source, int parc, char *parv[])
+mo_dline(struct Client *source, size_t parc, char *parv[])
 {
   if (!client_has_oper_flag(source, OPER_FLAG_DLINE))
   {
@@ -211,7 +211,7 @@ mo_dline(struct Client *source, int parc, char *parv[])
  *      - parv[4] = reason
  */
 static void
-ms_dline(struct Client *source, int parc, char *parv[])
+ms_dline(struct Client *source, size_t parc, char *parv[])
 {
   uintmax_t duration;
   if (io_parse_uintmax(parv[2], &duration) != IO_PARSE_OK)
