+++ modules/m_restart.c
@@ -33,7 +33,7 @@
  *      - parv[1] = server name
  */
 static void
-mo_restart(struct Client *source, int parc, char *parv[])
+mo_restart(struct Client *source, size_t parc, char *parv[])
 {
   if (!client_has_oper_flag(source, OPER_FLAG_RESTART))
   {
