+++ modules/m_svshost.c
@@ -32,7 +32,7 @@
  *      - parv[3] = host name
  */
 static void
-ms_svshost(struct Client *source, int parc, char *parv[])
+ms_svshost(struct Client *source, size_t parc, char *parv[])
 {
   if (!client_is_service(source) && !client_is_server(source))
     return;
